+++ Sources/Sentry/SentryHub.m
@@ -209,14 +209,9 @@ SentryHub () <SentryMetricsAPIDelegate>
         SentryClient *client = _client;
 
         if (client.options.diagnosticLevel == kSentryLevelDebug) {
-            NSData *sessionData = [NSJSONSerialization dataWithJSONObject:[session serialize]
-                                                                  options:0
-                                                                    error:nil];
-            NSString *sessionString = [[NSString alloc] initWithData:sessionData
-                                                            encoding:NSUTF8StringEncoding];
             [SentryLog
                 logWithMessage:[NSString stringWithFormat:@"Capturing session with status: %@",
-                                         sessionString]
+                                         [self createSessionDebugString:session]]
                       andLevel:kSentryLevelDebug];
         }
         [client captureSession:session];
@@ -630,6 +625,21 @@ SentryHub () <SentryMetricsAPIDelegate>
     }
 }
 
+/**
+ * Needed by hybrid SDKs as react-native to synchronously store an envelope to disk.
+ */
+- (void)storeEnvelope:(SentryEnvelope *)envelope
+{
+    SentryClient *client = _client;
+    if (client == nil) {
+        return;
+    }
+
+    // Envelopes are stored only when crash occurs. We should not start a new session when
+    // the app is about to crash.
+    [client storeEnvelope:[self updateSessionState:envelope startNewSession:NO]];
+}
+
 - (void)captureEnvelope:(SentryEnvelope *)envelope
 {
     SentryClient *client = _client;
@@ -637,10 +647,13 @@ SentryHub () <SentryMetricsAPIDelegate>
         return;
     }
 
-    [client captureEnvelope:[self updateSessionState:envelope]];
+    // If captured envelope cointains not handled errors, these are not going to crash the app and
+    // we should create new session.
+    [client captureEnvelope:[self updateSessionState:envelope startNewSession:YES]];
 }
 
 - (SentryEnvelope *)updateSessionState:(SentryEnvelope *)envelope
+                       startNewSession:(BOOL)startNewSession
 {
     BOOL handled = YES;
     if ([self envelopeContainsEventWithErrorOrHigher:envelope.items wasHandled:&handled]) {
@@ -654,9 +667,17 @@ SentryHub () <SentryMetricsAPIDelegate>
                 [currentSession
                     endSessionCrashedWithTimestamp:[SentryDependencyContainer.sharedInstance
                                                            .dateProvider date]];
-                // Setting _session to nil so startSession doesn't capture it again
-                _session = nil;
-                [self startSession];
+                if (_client.options.diagnosticLevel == kSentryLevelDebug) {
+                    [SentryLog
+                        logWithMessage:[NSString stringWithFormat:@"Ending session with status: %@",
+                                                 [self createSessionDebugString:currentSession]]
+                              andLevel:kSentryLevelDebug];
+                }
+                if (startNewSession) {
+                    // Setting _session to nil so startSession doesn't capture it again
+                    _session = nil;
+                    [self startSession];
+                }
             }
         }
 
@@ -713,6 +734,18 @@ SentryHub () <SentryMetricsAPIDelegate>
         SENTRY_LOG_DEBUG(@"The options `enableTimeToFullDisplay` is disabled.");
     }
 #endif // SENTRY_HAS_UIKIT
+}
+
+- (NSString *)createSessionDebugString:(SentrySession *)session
+{
+    if (session == nil) {
+        return @"Session is nil.";
+    }
+
+    NSData *sessionData = [NSJSONSerialization dataWithJSONObject:[session serialize]
+                                                          options:0
+                                                            error:nil];
+    return [[NSString alloc] initWithData:sessionData encoding:NSUTF8StringEncoding];
 }
 
 - (void)flush:(NSTimeInterval)timeout